+++ openhands/server/routes/settings.py
@@ -4,8 +4,8 @@ from fastapi.responses import JSONResponse
 from openhands.core.logger import openhands_logger as logger
 from openhands.server.settings import Settings
 from openhands.server.shared import config, openhands_config
-from openhands.storage.conversation_store import ConversationStore
-from openhands.storage.settings_store import SettingsStore
+from openhands.storage.conversation.conversation_store import ConversationStore
+from openhands.storage.settings.settings_store import SettingsStore
 from openhands.utils.import_utils import get_impl
 
 app = APIRouter(prefix='/api')
