+++ openhands/storage/conversation/file_conversation_store.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 
 from openhands.core.config.app_config import AppConfig
 from openhands.storage import get_file_store
-from openhands.storage.conversation_store import ConversationStore
+from openhands.storage.conversation.conversation_store import ConversationStore
 from openhands.server.data_models.conversation_metadata import ConversationMetadata
 from openhands.storage.files import FileStore
 from openhands.storage.locations import get_conversation_metadata_filename
