+++ tests/unit/test_file_settings_store.py
@@ -5,7 +5,7 @@ import pytest
 
 from openhands.core.config.app_config import AppConfig
 from openhands.server.settings import Settings
-from openhands.storage.file_settings_store import FileSettingsStore
+from openhands.storage.settings.file_settings_store import FileSettingsStore
 from openhands.storage.files import FileStore
 
 
@@ -15,18 +15,18 @@ def mock_file_store():
 
 
 @pytest.fixture
-def session_init_store(mock_file_store):
+def file_settings_store(mock_file_store):
     return FileSettingsStore(mock_file_store)
 
 
 @pytest.mark.asyncio
-async def test_load_nonexistent_data(session_init_store):
-    session_init_store.file_store.read.side_effect = FileNotFoundError()
-    assert await session_init_store.load() is None
+async def test_load_nonexistent_data(file_settings_store):
+    file_settings_store.file_store.read.side_effect = FileNotFoundError()
+    assert await file_settings_store.load() is None
 
 
 @pytest.mark.asyncio
-async def test_store_and_load_data(session_init_store):
+async def test_store_and_load_data(file_settings_store):
     # Test data
     init_data = Settings(
         language='python',
@@ -40,19 +40,19 @@ async def test_store_and_load_data(session_init_store):
     )
 
     # Store data
-    await session_init_store.store(init_data)
+    await file_settings_store.store(init_data)
 
     # Verify store called with correct JSON
     expected_json = json.dumps(init_data.__dict__)
-    session_init_store.file_store.write.assert_called_once_with(
+    file_settings_store.file_store.write.assert_called_once_with(
         'settings.json', expected_json
     )
 
     # Setup mock for load
-    session_init_store.file_store.read.return_value = expected_json
+    file_settings_store.file_store.read.return_value = expected_json
 
     # Load and verify data
-    loaded_data = await session_init_store.load()
+    loaded_data = await file_settings_store.load()
     assert loaded_data is not None
     assert loaded_data.language == init_data.language
     assert loaded_data.agent == init_data.agent
@@ -69,7 +69,7 @@ async def test_get_instance():
     config = AppConfig(file_store='local', file_store_path='/test/path')
 
     with patch(
-        'openhands.storage.file_settings_store.get_file_store'
+        'openhands.storage.settings.file_settings_store.get_file_store'
     ) as mock_get_store:
         mock_store = MagicMock(spec=FileStore)
         mock_get_store.return_value = mock_store
